pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "artgen"
version = "0.1.0"
description = "Batch AI art generation with interactive approval workflow"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["ai", "art", "generation", "gemini", "dall-e", "sprites"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "click>=8.0.0",
    "google-genai>=1.0.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "httpx>=0.26.0",
    "Pillow>=10.2.0",
    "rembg>=2.0.57",
    "pymatting>=1.1.12",
    "onnxruntime>=1.17.0",
    "scikit-image>=0.22.0",
    "pydantic>=2.6.0",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "rich>=13.0.0",
    "simpleeval>=0.9.13",
    "litellm>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "litellm>=1.0.0",
]

[project.scripts]
artgen = "backend.artgen:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "frontend*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.toml", "*.yaml", "*.html", "*.css", "*.js"]
